getting image metadata refactor
